+++ excel/src/robocorp/excel/__init__.py
@@ -9,7 +9,7 @@ in your dependencies' configuration.
 from robocorp.excel.excel import create_workbook, open_workbook
 from robocorp.excel.tables import Table
 
-__version__ = "0.4.3"
+__version__ = "0.4.4"
 __all__ = [
     "create_workbook",
     "open_workbook",