+++ log_pytest/src/robocorp/log_pytest/__init__.py
@@ -14,7 +14,7 @@ from robocorp.log.redirect import setup_stdout_logging
 
 from robocorp import log
 
-__version__ = "0.0.3"
+__version__ = "0.0.4"
 version_info = [int(x) for x in __version__.split(".")]